@@ -1,4 +1,4 @@
 /*!
 * TOTUM LOCALIZATION
 * */
-App.langs=App.langs||{},App.langs.en={locale:"en-EN",localeDatetimepicker:"en",dateFormat:"DD/MM/YY",dateTimeFormat:"DD/MM/YY HH:mm",timeDateFormatNoYear:"HH:mm DD/MM",filtersExtenders:App.commonFiltersExtenders,search_prepare_function:function(e,t){return e=e.toLowerCase(),t&&(t=t.toLowerCase()),[e,t]},css:{table:'.pcTable-container .loading-row td {background: url("/imgs/loading_en.png") repeat #fff;}'},modelMethods:{edit:"Editing",checkInsertRow:"Pre-addition",duplicate:"Duplicating",refresh_rows:"Rows refreshing",loadPage:"Page loaging",getTableData:"Table ingo loading",refresh:"Table data refreshing",checkEditRow:"Panel pre-calculating",saveEditRow:"Panel saving",save:"Field changing",click:"Button click",selectSourceTableAction:"Select source panel",add:"Row adding",getEditSelect:"Select loading",delete:"Deleting"},translates:{"Creator-tableEditButtons-default_action":"Action","Creator-tableEditButtons-on_duplicate":"Duplicate","Creator-tableEditButtons-row_format":"Row format","Creator-tableEditButtons-table_format":"Table format","Close context data":"<b>Close</b> context data","Open context data":"<b>Open</b> context data","PATH-TO-DOCUMENTATION":"https://docs.totum.online/",FormatShort:"Format",ApplyShort:"Apply",InvertShort:"Invert",CancelShort:"Cancel",ActionShort:"Action",SelectShort:"Select",__clock_shelve_panel:'<span className="clocks-na">For</span> <input type="number" step="1" value="10" className="form-control"/> <select className="form-control"><option  selected value="1">minutes</option><option value="2">hours</option><option value="3">days</option></select>',"Log is empty":"Log is empty. Enable logging and reload the page",treeAddTable:"Table",treeAddFolder:"Folder/Link","isCreatorSelector-NotCreatorView":"Switch off admin layer","isCreatorSelector-CommonView":"Switch off special view","isCreatorSelector-MobileView":"Switch to desktop view",mobileToDesctopWarning:"This display type is only for PCs with a small screen. Do not turn it on if you have a mobile device such as a phone or *PAD",mobileToDesctopUserWarning:"We detected the page type automatically.\n\nIf we made a mistake, you can switch the mobile/desktop view manually.\n\nYou have to be sure of the action you're performing!\n\nIf you switch to desktop view on your mobile device, the page will not work!","versions(%s)":"versions (%s)","Single editing":"Editing","bugFinder-warning":"Testing is in non-save mode and safe for the database, but may invoke third-party services and send emails if those functions are involved in the code chain of the page being opened."}};
+App.langs=App.langs||{},App.langs.en={locale:"en-EN",localeDatetimepicker:"en",dateFormat:"DD/MM/YY",dateTimeFormat:"DD/MM/YY HH:mm",timeDateFormatNoYear:"HH:mm DD/MM",filtersExtenders:App.commonFiltersExtenders,search_prepare_function:function(e,t){return e=e.toLowerCase(),t&&(t=t.toLowerCase()),[e,t]},css:{table:'.pcTable-container .loading-row td {background: url("/imgs/loading_en.png") repeat #fff;}'},modelMethods:{edit:"Editing",checkInsertRow:"Pre-addition",duplicate:"Duplicating",refresh_rows:"Rows refreshing",loadPage:"Page loaging",getTableData:"Table ingo loading",refresh:"Table data refreshing",checkEditRow:"Panel pre-calculating",saveEditRow:"Panel saving",save:"Field changing",click:"Button click",selectSourceTableAction:"Select source panel",add:"Row adding",getEditSelect:"Select loading",delete:"Deleting"},translates:{"Creator-tableEditButtons-default_action":"Action","Creator-tableEditButtons-on_duplicate":"Duplicate","Creator-tableEditButtons-row_format":"Row format","Creator-tableEditButtons-table_format":"Table format","Close context data":"<b>Close</b> context data","Open context data":"<b>Open</b> context data","PATH-TO-DOCUMENTATION":"https://docs.totum.online/",Forms:"PRO Forms","Add form":"PRO Add form","Click hear to unlock":"Click here to unlock",FormatShort:"Format",ApplyShort:"Apply",InvertShort:"Invert",CancelShort:"Cancel",ActionShort:"Action",SelectShort:"Select",__clock_shelve_panel:'<span className="clocks-na">For</span> <input type="number" step="1" value="10" className="form-control"/> <select className="form-control"><option  selected value="1">minutes</option><option value="2">hours</option><option value="3">days</option></select>',"Log is empty":"Log is empty. Enable logging and reload the page",treeAddTable:"Table",treeAddFolder:"Folder/Link","isCreatorSelector-NotCreatorView":"Switch off admin layer","isCreatorSelector-CommonView":"Switch off special view","isCreatorSelector-MobileView":"Switch to desktop view",mobileToDesctopWarning:"This display type is only for PCs with a small screen. Do not turn it on if you have a mobile device such as a phone or *PAD",mobileToDesctopUserWarning:"We detected the page type automatically.\n\nIf we made a mistake, you can switch the mobile/desktop view manually.\n\nYou have to be sure of the action you're performing!\n\nIf you switch to desktop view on your mobile device, the page will not work!","versions(%s)":"versions (%s)","Single editing":"Editing","bugFinder-warning":"Testing is in non-save mode and safe for the database, but may invoke third-party services and send emails if those functions are involved in the code chain of the page being opened.",userNotFoundError:"The user was logged out because their account was disabled or the maximum number of simultaneous connections was exceeded"}};
