@@ -5,7 +5,6 @@ namespace totum\commands;
 
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use totum\config\Conf;
 
@@ -17,19 +16,12 @@ class CleanSchemasTmpTables extends Command
             ->setDescription('Clean tmp_tables in schemas. For multi install. Set in crontab one time in 10 minutes.');
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $Conf = new Conf();
-        $sql = $Conf->getSql(withSchema: false);
-        $plus24 = date_create();
-        $plus24->modify('-24 hours');
-        $minus10 = date_create();
-        $minus10->modify('-10 minutes');
-
         foreach (array_unique(array_values(Conf::getSchemas())) as $schema) {
-            $sql->exec('delete from "'.$schema.'"._tmp_tables where touched<\'' . $plus24->format('Y-m-d H:i') . '\'');
-            $sql->exec('delete from "'.$schema.'"._tmp_tables where table_name SIMILAR TO \'\_%\' AND touched<\'' . $minus10->format('Y-m-d H:i') . '\'');
-            $sql->exec('VACUUM "'.$schema.'"._tmp_tables');
+            CleanSchemaTmpTables::doSqlWorks($schema, $Conf);
         }
+        return 0;
     }
 }
