@@ -7,11 +7,8 @@ use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Question\ConfirmationQuestion;
-use totum\common\Auth;
 use totum\common\configs\MultiTrait;
 use totum\common\errorException;
-use totum\common\Totum;
 use totum\config\Conf;
 
 class Vacuum extends Command
@@ -28,7 +25,7 @@ class Vacuum extends Command
         $this->addOption('analyze', 'a', InputOption::VALUE_NONE, 'With analyze');
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         if (!class_exists(Conf::class)) {
             $output->writeln('ERROR: config class not found');
