@@ -1255,6 +1255,9 @@ class CalculateAction extends Calculate
             if ($filters) {
                 $cripted = Crypt::getCrypted(json_encode($filters, JSON_UNESCAPED_UNICODE));
                 $q_params['f'] = $cripted;
+                if (!empty($q_params['b'])) {
+                    $q_params['efl'] = $q_params['f'];
+                }
             }
         }
 
@@ -1329,7 +1332,7 @@ class CalculateAction extends Calculate
 
     protected function funcInsert($params)
     {
-        if ($params = $this->getParamsArray($params, ['field', 'cycle'], ['field'])) {
+        if ($params = $this->getParamsArray($params, ['field'], ['field'])) {
             $addedIds = [];
             $funcSet = function ($params) use (&$addedIds) {
                 $table = $this->getSourceTable($params);
@@ -1362,7 +1365,7 @@ class CalculateAction extends Calculate
 
 
             if (!empty($params['cycle'])) {
-                $cycleIds = $params['cycle'];
+                $cycleIds = (array)$params['cycle'];
                 foreach ($cycleIds as $cycleId) {
                     $params['cycle'] = $cycleId;
                     $funcSet($params);
