@@ -662,38 +662,6 @@ class ReadTableActions extends Actions
 
 
         switch ($pageViewType = $this->getPageViewType()) {
-            case 'tree':
-                $result += ['chdata' => []];
-
-                if (!is_array($this->post) || !key_exists('tree', $this->post)) {
-                    throw new errorException($this->translate('The tree index is not passed'));
-                }
-                $treeIndex = json_decode($this->post['tree'], true);
-                $result['chdata'] = array_merge(
-                    $result['chdata'],
-                    $this->getResultTree(
-                        function ($k, $v) use ($treeIndex) {
-                            if (key_exists($k, $treeIndex)) {
-                                if ($treeIndex[$k]) {
-                                    return 'loaded';
-                                } else {
-                                    return 'child';
-                                }
-                            }
-                        },
-                        ['']
-                    )
-                );
-
-                if ($this->isPagingView('tree') && $ids = json_decode($this->post['ids'], true)) {
-                    $result['chdata']['rows'] = $this->Table->getSortedFilteredRows('web', 'web', $ids)['rows'];
-                }
-
-                $pageIds = array_column($result['chdata']['rows'] ?? [], 'id');
-                $result['chdata']['params'] = $this->addValuesAndFormatsOfParams($this->Table->getTbl()['params'],
-                    $pageIds)['params'];
-                $result['chdata']['f'] = $this->getTableFormat($pageIds);
-                break;
             default:
 
                 $result += ['chdata' => $this->getTableClientData(
@@ -706,19 +674,12 @@ class ReadTableActions extends Actions
                     false
                 )];
 
-                switch ($pageViewType) {
-                    case 'panels':
-                        if ($this->Table->getTableRow()['with_order_field']) {
-                            $result['chdata']['nsorted_ids'] = array_column($result['chdata']['rows'], 'id');
-                        }
-                        break;
-                    default:
-                        if ($this->isPagingView()) {
-                            $params = $this->Table->filtersParamsForLoadRows('web');
-                            $result['allCount'] = $params === false ? 0 : $this->Table->countByParams($params);
-                        }
-                        break;
+
+                if ($this->isPagingView()) {
+                    $params = $this->Table->filtersParamsForLoadRows('web');
+                    $result['allCount'] = $params === false ? 0 : $this->Table->countByParams($params);
                 }
+                break;
         }
 
 
@@ -1260,38 +1221,6 @@ table tr td.title{font-weight: bold}', 'html' => '{table}'];
         }
 
         switch ($this->getPageViewType()) {
-            case 'panels':
-                $result['viewType'] = 'panels';
-                $result['kanban'] = $this->getKanbanData($fields);
-
-                foreach ($result['fields'] as $k => $v) {
-                    if ($v['category'] === 'column' && !in_array($k, $fields)) {
-                        unset($result['fields'][$k]);
-                    }
-                }
-
-                $result = array_merge(
-                    $result,
-                    $this->getTableClientData(
-                        0,
-                        $this->isPagingView() ? 0 : null,
-                        false,
-                        $fields
-                    )
-                );
-                break;
-            case 'tree':
-                $tree = $this->Table->getFields()['tree'];
-                $result = array_merge(
-                    $result,
-                    $this->getTreeTopLevel($tree['treeViewLoad'] ?? null,
-                        $tree['treeViewOpen'] ?? null)
-                );
-                break;
-            case 'commonByCount':
-                /*For off button on table head*/
-                $result['panels'] = 'off';
-            // no break
             default:
                 $result = array_merge($result, $this->getTableClientData(0, $this->isPagingView() ? 0 : null, false));
 
@@ -1325,46 +1254,6 @@ table tr td.title{font-weight: bold}', 'html' => '{table}'];
 
     protected function getPageViewType(): string
     {
-        if (($this->post['restoreView'] ?? false) || ($this->User->isCreator() && ($this->Cookies['ttm__commonTableView'] ?? false))) {
-            $this->creatorCommonView = true;
-            return 'common';
-        }
-
-        if ($this->Request->getQueryParams()['iframe'] ?? false) ; elseif (($panelViewSettings = ($this->Table->getTableRow()['panels_view'] ?? null))
-        ) {
-            if (($this->post['panelsView'] ?? false) === 'true') {
-                return 'panels';
-            } elseif (empty($this->post)) {
-                $params = $this->Table->filtersParamsForLoadRows('web');
-                $allCount = $params === false ? 0 : $this->Table->countByParams($params);
-                if ($allCount <= ($panelViewSettings['panels_max_count'] ?? 100)) {
-                    $checkCookies = function () use ($panelViewSettings) {
-                        $name = $this->getPanelsCookieName()[0];
-                        if (key_exists($name, $_COOKIE)) {
-                            return $_COOKIE[$name] === '1';
-                        }
-                        return $panelViewSettings['panels_view_first'];
-                    };
-
-                    if ($panelViewSettings['state'] === 'panel'
-                        || (
-                            $panelViewSettings['state'] === 'both'
-                            && $checkCookies()
-                        )) {
-                        return 'panels';
-                    }
-                } else {
-                    return 'commonByCount';
-                }
-            }
-        }
-        if (($tree = $this->Table->getFields()['tree'] ?? null)
-            && $tree['category'] === 'column'
-            && $tree['type'] === 'tree'
-            && !empty($tree['treeViewType'])) {
-            return 'tree';
-        }
-
         return 'common';
     }
 
@@ -2044,12 +1933,12 @@ table tr td.title{font-weight: bold}', 'html' => '{table}'];
             );
 
         } else {
-            $tableRow['description'] = preg_replace('`\s*<admin>.*?</admin>\s*`su', '', $tableRow['description']);
+            $tableRow['description'] = preg_replace('`\s*<admin>.*?</admin>\s*`su', '', $tableRow['description'] ?? '');
         }
         $_tableRow = array_intersect_key($tableRow, array_flip($fields));
         foreach (Totum::TABLE_CODE_PARAMS as $name) {
             if (key_exists($name, $_tableRow)) {
-                $_tableRow[$name] = trim($_tableRow[$name]);
+                $_tableRow[$name] = trim($_tableRow[$name] ?? '');
                 $_tableRow[$name] = !!preg_match('/^\s*[a-z0-9]*\=\:\s*[^\s]+/mu', $_tableRow[$name]);
             }
         }
@@ -2082,7 +1971,7 @@ table tr td.title{font-weight: bold}', 'html' => '{table}'];
             }
 
         }
-        $_tableRow['description'] = preg_replace('#<script(.*?)>(.*?)</script>#is', '', $_tableRow['description']);
+        $_tableRow['description'] = preg_replace('#<script(.*?)>(.*?)</script>#is', '', $_tableRow['description'] ?? '');
         $_tableRow['__withPDF'] = $this->isTableServiceOn('pdf') && !$this->isServicesBlocked && !$this->Totum->getConfig()->isTechTable($this->Table->getTableRow()['name']);
         $_tableRow['__xlsx'] = $this->isTableServiceOn('xlsx') && !$this->isServicesBlocked && !$this->Totum->getConfig()->isTechTable($this->Table->getTableRow()['name']);
         $_tableRow['__xlsx_import'] = is_a($this, WriteTableActions::class) && $this->isTableServiceOn('xlsximport') && !$this->Totum->getConfig()->isTechTable($this->Table->getTableRow()['name']) && !$this->isServicesBlocked && key_exists($this->Totum->getTableRow('ttm__prepared_data_import')['id'], $this->User->getTables());
@@ -2266,9 +2155,6 @@ table tr td.title{font-weight: bold}', 'html' => '{table}'];
                     unset($changedIds['params']);
 
                     $selectOrFormatColumns['id'] = true;
-                    if ($this->getPageViewType() === 'tree') {
-                        $selectOrFormatColumns['n'] = true;
-                    }
 
                     $rows = $this->Table->getTbl()['rows'];
                     foreach ($pageIds as $id) {
@@ -2303,15 +2189,6 @@ table tr td.title{font-weight: bold}', 'html' => '{table}'];
                 if ($this->Table->getChangeIds()['reordered']) {
                     $return['chdata']['order'] = array_column($return['chdata']['rows'], 'id');
                 }
-            }
-            if ($this->getPageViewType() === 'tree' && $this->Table->getFields()['tree']['treeViewType'] === 'self') {
-                $return['chdata']['tree'] = $this->getResultTree(
-                    function ($k, $v) {
-                        return 'child';
-                    },
-                    [''],
-                    true
-                )['tree'];
             }
 
             $Log = $this->Table->calcLog(['name' => 'SELECTS AND FORMATS']);
@@ -2699,7 +2576,7 @@ table tr td.title{font-weight: bold}', 'html' => '{table}'];
         }
 
         foreach ($data as &$v) {
-            $v = json_decode($v, true);
+            $v = json_decode($v??'', true);
         }
         if (is_array($data['tables'])) {
             if (in_array('*ALL*', $data['tables'])) {
