@@ -136,7 +136,7 @@ class TableController extends interfaceController
             if ($links = $this->Totum->getInterfaceDatas()) {
                 $result['interfaceDatas'] = $links;
             }
-            if ($tableChanged && $method !== 'refresh') {
+            if ($tableChanged && $method !== 'refresh' && $this->Table->getTableRow()['actual'] != 'disable') {
                 $tableChanged['username'] = $this->Totum->getNamedModel(UserV::class)->getFio($tableChanged['user'],
                     true);
                 $result['tableChanged'] = $tableChanged;
@@ -791,7 +791,7 @@ class TableController extends interfaceController
                     case null:
                         $this->__addAnswerVar(
                             'html',
-                            preg_replace('#<script(.*?)>(.*?)</script>#is', '', $branchData['html'])
+                            preg_replace('#<script(.*?)>(.*?)</script>#is', '', $branchData['html']??'')
                         );
                         break;
                     case 'anchor':
@@ -957,7 +957,7 @@ class TableController extends interfaceController
                 } elseif (!($field = $this->Table->getFields()[$fieldName])) {
                     $error = $this->translate('The file field was not found');
                 } else {
-                    $filepath = File::getFilePath($filename, $this->Config, $field);
+                    $filepath = File::getFilePath($filename, $this->Config);
                 }
             }
             if (!empty($filepath)) {
