@@ -3,22 +3,22 @@
     <script>App = {}</script>
     <link rel="stylesheet"
           type="text/css"
-          href="/css/libs.css?v=ead40ab">
-    <script src="/js/libs.js?v=665342a"></script>
+          href="/css/libs.css?v=3710809">
+    <script src="/js/libs.js?v=b101ece"></script>
     <link rel="stylesheet"
           type="text/css"
-          href="/css/main.css?v=1f0a57c">
+          href="/css/main.css?v=de85130">
 
     <?php
     if ($isCreatorView ?? null) { ?>
-        <script src="/js/functions.js?v=4718f08"></script>
+        <script src="/js/functions.js?v=57abc29"></script>
         <?php
         echo '<script>App.functions=App.functions.concat(' . $this->Config->getExtFunctionsTemplates() . ')</script>';
         ?>
         <?php
     } ?>
 
-    <script src="/js/main.js?v=5280e03"></script>
+    <script src="/js/main.js?v=707fff3"></script>
     <script src="/js/i18n/<?= $this->Config->getLang() ?>.js?10"></script>
     <script>App.lang = App.langs["<?= $this->Config->getLang() ?>"]</script>
 
